+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file(see http://k
 ## [Unreleased]
 ### Added
 - fs: expand '~' to home directory when use realpath and absolute.
+- ini: user can retrieve error position when syntax error.
+- json: user can retrieve error position when syntax error.
 
 ## [1.1.0] - 2016-07-22
 ### Added